Homework 4 + изменения по предыдущему code review
